+++ service/src/common/enums/providers.ts
@@ -1,2 +1,1 @@
 export const NOTIFICATIONS_PROVIDER = 'NOTIFICATIONS';
-export const TEMPLATE_PROVIDER = 'TEMPLATES';
