--- service/src/services/domain/builders/index.ts
@@ -1,22 +1,0 @@
-export * from './space/space.admin.community.application.received.notification.builder';
-export * from './space/space.communication.update.notification.builder';
-export * from './space/space.lead.communication.message.direct.notification.builder';
-export * from './space/space.admin.community.new.member.notification.builder';
-export * from './space/space.collaboration.callout.contribution.notification.builder';
-export * from './space/space.collaboration.callout.post.contribution.comment.notification.builder';
-export * from './space/space.collaboration.callout.published.notification.builder';
-export * from './organization/organization.message.sender.notification.builder';
-export * from './organization/organization.message.recipient.notification.builder';
-export * from './organization/organization.mention.notification.builder';
-export * from './user/user.space.community.joined.notification.builder';
-export * from './user/user.space.community.application.submitted.notification.builder';
-export * from './user/user.space.community.invitation.received.notification.builder';
-export * from './user/space.community.invitation.platform.created.notification.builder';
-export * from './user/user.message.recipient.notification.builder';
-export * from './user/user.message.sender.notification.builder';
-export * from './user/user.mention.notification.builder';
-export * from './user/user.comment.reply.notification.builder';
-export * from './user/user.sign.up.welcome.notification.builder';
-export * from './platform/platform.admin.user.profile.removed.notification.builder';
-export * from './platform/platform.forum.discussion.comment.notification.builder';
-export * from './platform/platform.forum.discussion.created.notification.builder';
