--- service/src/services/domain/builders/notification.builder.interface.ts
@@ -1,9 +1,0 @@
-import { BaseEventPayload } from '@alkemio/notifications-lib';
-import { PlatformUser, User } from '../../../core/models/user';
-import { BaseEmailPayload } from '@src/common/email-template-payload';
-export interface INotificationBuilder {
-  createEmailTemplatePayload(
-    eventPayload: BaseEventPayload,
-    recipient: User | PlatformUser
-  ): BaseEmailPayload;
-}
