--- service/src/services/domain/builders/organization/organization.mention.notification.builder.ts
@@ -1,46 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { PlatformUser, User } from '@core/models';
-import { INotificationBuilder } from '@src/services/domain/builders/notification.builder.interface';
-import { createUserNotificationPreferencesURL } from '@src/core/util/createNotificationUrl';
-import { CommunicationOrganizationMentionEmailPayload } from '@common/email-template-payload';
-import { NotificationEventPayloadOrganizationMessageRoom } from '@alkemio/notifications-lib';
-import { convertMarkdownToText } from '@src/utils/markdown-to-text.util';
-@Injectable()
-export class OrganizationMentionNotificationBuilder
-  implements INotificationBuilder
-{
-  constructor() {}
-
-  public createEmailTemplatePayload(
-    eventPayload: NotificationEventPayloadOrganizationMessageRoom,
-    recipient: User | PlatformUser
-  ): CommunicationOrganizationMentionEmailPayload {
-    const notificationPreferenceURL =
-      createUserNotificationPreferencesURL(recipient);
-
-    const htmlComment: string = convertMarkdownToText(eventPayload.comment);
-
-    return {
-      commentSender: {
-        displayName: eventPayload.triggeredBy.profile.displayName,
-        firstName: eventPayload.triggeredBy.firstName,
-      },
-      recipient: {
-        firstName: recipient.firstName,
-        email: recipient.email,
-        notificationPreferences: notificationPreferenceURL,
-      },
-      comment: htmlComment,
-      platform: {
-        url: eventPayload.platform.url,
-      },
-      mentionedOrganization: {
-        displayName: eventPayload.organization.profile.displayName,
-      },
-      commentOrigin: {
-        url: eventPayload.commentOrigin.url,
-        displayName: eventPayload.commentOrigin.displayName,
-      },
-    };
-  }
-}
