--- service/src/services/domain/builders/platform/platform.admin.space.created.notification.builder.ts
@@ -1,41 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { INotificationBuilder } from '../notification.builder.interface';
-import { User } from '@src/core/models';
-import { SpaceCreatedEmailPayload } from '@src/common/email-template-payload';
-import { NotificationEventPayloadPlatformSpaceCreated } from '@alkemio/notifications-lib';
-import { createUserNotificationPreferencesURL } from '@src/core/util/createNotificationUrl';
-
-@Injectable()
-export class PlatformAdminSpaceCreatedNotificationBuilder
-  implements INotificationBuilder
-{
-  constructor() {}
-
-  createEmailTemplatePayload(
-    eventPayload: NotificationEventPayloadPlatformSpaceCreated,
-    recipient: User
-  ): SpaceCreatedEmailPayload {
-    const notificationPreferenceURL =
-      createUserNotificationPreferencesURL(recipient);
-
-    return {
-      recipient: {
-        firstName: recipient.firstName,
-        email: recipient.email,
-        notificationPreferences: notificationPreferenceURL,
-      },
-      space: {
-        displayName: eventPayload.space.profile.displayName,
-        level: eventPayload.space.level,
-        url: eventPayload.space.profile.url,
-      },
-      sender: eventPayload.sender,
-      dateCreated: new Date(eventPayload.created).toLocaleString('en-GB', {
-        timeZone: 'UTC',
-      }),
-      platform: {
-        url: eventPayload.platform.url,
-      },
-    };
-  }
-}
