--- service/src/services/domain/builders/space/space.admin.community.new.member.notification.builder.ts
@@ -1,49 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { INotificationBuilder } from '../notification.builder.interface';
-import { User } from '@core/models';
-import { createUserNotificationPreferencesURL } from '@src/core/util/createNotificationUrl';
-import { CommunityNewMemberEmailPayload } from '@common/email-template-payload';
-import { NotificationEventPayloadSpaceCommunityContributor } from '@alkemio/notifications-lib';
-import { RoleSetContributorType } from '@src/generated/alkemio-schema';
-
-@Injectable()
-export class SpaceAdminCommunityNewMemberNotificationBuilder
-  implements INotificationBuilder
-{
-  constructor() {}
-
-  public createEmailTemplatePayload(
-    eventPayload: NotificationEventPayloadSpaceCommunityContributor,
-    recipient: User
-  ): CommunityNewMemberEmailPayload {
-    const notificationPreferenceURL =
-      createUserNotificationPreferencesURL(recipient);
-
-    const newMember = eventPayload.contributor;
-    const typeName =
-      newMember.type === RoleSetContributorType.Virtual
-        ? 'virtual contributor'
-        : newMember.type;
-
-    return {
-      member: {
-        name: newMember.profile.displayName,
-        profile: newMember.profile.url,
-        type: typeName,
-      },
-      recipient: {
-        firstName: recipient.firstName,
-        email: recipient.email,
-        notificationPreferences: notificationPreferenceURL,
-      },
-      space: {
-        displayName: eventPayload.space.profile.displayName,
-        level: eventPayload.space.level,
-        url: eventPayload.space.profile.url,
-      },
-      platform: {
-        url: eventPayload.platform.url,
-      },
-    };
-  }
-}
