--- service/src/services/domain/builders/space/space.collaboration.callout.contribution.notification.builder.ts
@@ -1,59 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { INotificationBuilder } from '../notification.builder.interface';
-import { User } from '@core/models';
-import { createUserNotificationPreferencesURL } from '@src/core/util/createNotificationUrl';
-import { CollaborationPostCreatedEmailPayload } from '@common/email-template-payload';
-import { NotificationEventPayloadSpaceCollaborationCallout } from '@alkemio/notifications-lib';
-import { EventPayloadNotProvidedException } from '@src/common/exceptions/event.payload.not.provided.exception';
-import { LogContext } from '@src/common/enums';
-
-@Injectable()
-export class SpaceCollaborationCalloutContributionNotificationBuilder
-  implements INotificationBuilder
-{
-  constructor() {}
-
-  createEmailTemplatePayload(
-    eventPayload: NotificationEventPayloadSpaceCollaborationCallout,
-    recipient: User
-  ): CollaborationPostCreatedEmailPayload {
-    const notificationPreferenceURL =
-      createUserNotificationPreferencesURL(recipient);
-
-    const contribution = eventPayload.callout.contribution;
-    if (!contribution) {
-      throw new EventPayloadNotProvidedException(
-        'Contribution not found',
-        LogContext.NOTIFICATION_BUILDER
-      );
-    }
-    return {
-      createdBy: {
-        firstName: eventPayload.triggeredBy.firstName,
-        email: eventPayload.triggeredBy.email,
-      },
-      callout: {
-        displayName: eventPayload.callout.framing.displayName,
-        url: eventPayload.callout.framing.url,
-      },
-      contribution: {
-        displayName: contribution.displayName,
-        url: contribution.url,
-        type: contribution.type,
-      },
-      recipient: {
-        firstName: recipient.firstName,
-        email: recipient.email,
-        notificationPreferences: notificationPreferenceURL,
-      },
-      space: {
-        displayName: eventPayload.space.profile.displayName,
-        level: eventPayload.space.level,
-        url: eventPayload.space.profile.url,
-      },
-      platform: {
-        url: eventPayload.platform.url,
-      },
-    };
-  }
-}
