--- service/src/services/domain/builders/space/space.collaboration.callout.post.contribution.comment.notification.builder.ts
@@ -1,58 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { INotificationBuilder } from '../notification.builder.interface';
-import { NotificationEventPayloadSpaceCollaborationCallout } from '@alkemio/notifications-lib';
-import { CollaborationPostCommentEmailPayload } from '@common/email-template-payload';
-import { createUserNotificationPreferencesURL } from '@src/core/util/createNotificationUrl';
-import { User } from '@core/models';
-import { EventPayloadNotProvidedException } from '@src/common/exceptions/event.payload.not.provided.exception';
-import { LogContext } from '@src/common/enums/logging.context';
-@Injectable()
-export class SpaceCollaborationCalloutPostContributionCommentNotificationBuilder
-  implements INotificationBuilder
-{
-  constructor() {}
-
-  createEmailTemplatePayload(
-    eventPayload: NotificationEventPayloadSpaceCollaborationCallout,
-    recipient: User
-  ): CollaborationPostCommentEmailPayload {
-    const notificationPreferenceURL =
-      createUserNotificationPreferencesURL(recipient);
-    const callout = eventPayload.callout;
-
-    const contribution = eventPayload.callout.contribution;
-    if (!contribution) {
-      throw new EventPayloadNotProvidedException(
-        'Contribution not found',
-        LogContext.NOTIFICATION_BUILDER
-      );
-    }
-    return {
-      callout: {
-        displayName: callout.framing.displayName,
-        url: callout.framing.url,
-      },
-      post: {
-        displayName: contribution.displayName,
-        url: contribution.url,
-      },
-      recipient: {
-        firstName: recipient.firstName,
-        email: recipient.email,
-        notificationPreferences: notificationPreferenceURL,
-      },
-      createdBy: {
-        firstName: eventPayload.triggeredBy.firstName,
-        email: eventPayload.triggeredBy.email,
-      },
-      space: {
-        displayName: eventPayload.space.profile.displayName,
-        level: eventPayload.space.level,
-        url: eventPayload.space.profile.url,
-      },
-      platform: {
-        url: eventPayload.platform.url,
-      },
-    };
-  }
-}
