--- service/src/services/domain/builders/space/space.collaboration.callout.published.notification.builder.ts
@@ -1,49 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { INotificationBuilder } from '../notification.builder.interface';
-import { User } from '@core/models';
-import { createUserNotificationPreferencesURL } from '@src/core/util/createNotificationUrl';
-import { NotificationEventPayloadSpaceCollaborationCallout } from '@alkemio/notifications-lib';
-import { CollaborationCalloutPublishedEmailPayload } from '@common/email-template-payload';
-import { normalizeCalloutType } from '@src/utils/callout.util';
-
-@Injectable()
-export class SpaceCollaborationCalloutPublishedNotificationBuilder
-  implements INotificationBuilder
-{
-  constructor() {}
-
-  createEmailTemplatePayload(
-    eventPayload: NotificationEventPayloadSpaceCollaborationCallout,
-    recipient: User
-  ): CollaborationCalloutPublishedEmailPayload {
-    const notificationPreferenceURL =
-      createUserNotificationPreferencesURL(recipient);
-
-    const framing = eventPayload.callout.framing;
-
-    const result: CollaborationCalloutPublishedEmailPayload = {
-      recipient: {
-        firstName: recipient.firstName,
-        email: recipient.email,
-        notificationPreferences: notificationPreferenceURL,
-      },
-      publishedBy: {
-        firstName: eventPayload.triggeredBy.profile.displayName,
-      },
-      callout: {
-        displayName: framing.displayName,
-        url: framing.url,
-        type: normalizeCalloutType(framing.type),
-      },
-      space: {
-        displayName: eventPayload.space.profile.displayName,
-        level: eventPayload.space.level,
-        url: eventPayload.space.profile.url,
-      },
-      platform: {
-        url: eventPayload.platform.url,
-      },
-    };
-    return result;
-  }
-}
