--- service/src/services/domain/builders/user/user.space.community.application.submitted.notification.builder.ts
@@ -1,47 +1,0 @@
-import { User } from '@core/models';
-import { createUserNotificationPreferencesURL } from '@src/core/util/createNotificationUrl';
-import { CommunityApplicationCreatedEmailPayload } from '@common/email-template-payload';
-import { Injectable } from '@nestjs/common';
-import { INotificationBuilder } from '../notification.builder.interface';
-import { NotificationEventPayloadSpaceCommunityApplication } from '@alkemio/notifications-lib';
-
-@Injectable()
-export class UserSpaceCommunityApplicationSubmittedNotificationBuilder
-  implements INotificationBuilder
-{
-  constructor() {}
-
-  public createEmailTemplatePayload(
-    eventPayload: NotificationEventPayloadSpaceCommunityApplication,
-    recipient: User
-  ): CommunityApplicationCreatedEmailPayload {
-    const isLevel0Space = eventPayload.space.level === '0';
-    const spaceType = isLevel0Space ? 'space' : 'subspace';
-
-    const notificationPreferenceURL =
-      createUserNotificationPreferencesURL(recipient);
-    return {
-      applicant: {
-        firstName: eventPayload.triggeredBy.firstName,
-        name: eventPayload.triggeredBy.profile.displayName,
-        email: eventPayload.triggeredBy.email,
-        profile: eventPayload.triggeredBy.profile.url,
-      },
-      spaceAdminURL: eventPayload.space.adminURL,
-      recipient: {
-        firstName: recipient.firstName,
-        email: recipient.email,
-        notificationPreferences: notificationPreferenceURL,
-      },
-      space: {
-        displayName: eventPayload.space.profile.displayName,
-        level: eventPayload.space.level,
-        url: eventPayload.space.profile.url,
-        type: spaceType,
-      },
-      platform: {
-        url: eventPayload.platform.url,
-      },
-    };
-  }
-}
