--- service/src/services/domain/builders/user/user.space.community.invitation.received.notification.builder.ts
@@ -1,56 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { INotificationBuilder } from '../notification.builder.interface';
-import { PlatformUser, User } from '@core/models';
-import { createUserNotificationPreferencesURL } from '@src/core/util/createNotificationUrl';
-import { CommunityInvitationCreatedEmailPayload } from '@common/email-template-payload';
-import { ConfigService } from '@nestjs/config';
-import { ConfigurationTypes } from '@src/common/enums';
-import { NotificationEventPayloadSpaceCommunityInvitation } from '@alkemio/notifications-lib';
-
-@Injectable()
-export class UserSpaceCommunityInvitationReceivedNotificationBuilder
-  implements INotificationBuilder
-{
-  invitationsPath: string;
-  constructor(private readonly configService: ConfigService) {
-    this.invitationsPath = this.configService.get(
-      ConfigurationTypes.ALKEMIO
-    )?.webclient_invitations_path;
-  }
-
-  public createEmailTemplatePayload(
-    eventPayload: NotificationEventPayloadSpaceCommunityInvitation,
-    recipient: User | PlatformUser
-  ): CommunityInvitationCreatedEmailPayload {
-    const notificationPreferenceURL =
-      createUserNotificationPreferencesURL(recipient);
-    const invitationsURL = `${eventPayload.platform.url.replace(/\/+$/, '')}${
-      this.invitationsPath
-    }`;
-
-    return {
-      inviter: {
-        firstName: eventPayload.triggeredBy.firstName,
-        name: eventPayload.triggeredBy.profile.displayName,
-        email: eventPayload.triggeredBy.email,
-        profile: eventPayload.triggeredBy.profile.url,
-      },
-      spaceAdminURL: eventPayload.space.adminURL,
-      recipient: {
-        firstName: recipient.firstName,
-        email: recipient.email,
-        notificationPreferences: notificationPreferenceURL,
-      },
-      welcomeMessage: eventPayload.welcomeMessage,
-      space: {
-        displayName: eventPayload.space.profile.displayName,
-        level: eventPayload.space.level,
-        url: eventPayload.space.profile.url,
-      },
-      platform: {
-        url: eventPayload.platform.url,
-      },
-      invitationsURL,
-    };
-  }
-}
