+++ service/src/services/notification/notifIcation.module.ts
@@ -1,10 +1,10 @@
-import { ConfigModule } from '@nestjs/config';
 import { Global, Module } from '@nestjs/common';
 import { NotificationEmailPayloadBuilderService } from './notification.email.payload.builder.service';
 import { NotificationService } from './notification.service';
+import { NotifmeModule } from '../notifme/notifme.module';
 @Global()
 @Module({
-  imports: [ConfigModule],
+  imports: [NotifmeModule],
   providers: [NotificationService, NotificationEmailPayloadBuilderService],
   exports: [NotificationService, NotificationEmailPayloadBuilderService],
 })
