+++ service/src/services/notification/notification.service.spec.ts
@@ -44,7 +44,7 @@ import { PlatformAdminSpaceCreatedNotificationBuilder } from '../builders/platfo
 import { NotificationEventPayloadSpaceCommunityApplication } from '@alkemio/notifications-lib';
 import { SpaceCommunicationMessageDirectSenderNotificationBuilder } from '../builders/space/space.communication.message.direct.sender.notification.builder';
 import { PlatformAdminUserProfileCreatedNotificationBuilder } from '../builders/platform/platform.admin.user.profile.created.notification.builder';
-import { NotificationTemplateBuilder } from '@src/services/external/notifme';
+import { NotificationTemplateBuilder } from '@src/services/notifme';
 import { SpaceAdminCollaborationCalloutContributionNotificationBuilder } from '../builders/space/space.admin.collaboration.callout.contribution.notification.builder';
 
 const testData = {
